+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.10.9 (unreleased)
+
+- Handle path to templates starting with "./"
+- Fix loop and macro context overlaps
+
+
 ## 0.10.8 (2017-06-24)
 
 - Update chrono